requirements.txt
together
pandas
termcolor
pexpect
scipy
ipdb
absl-py
diskcache
requests
python-dotenv
nltk
sacrebleu